Change Project Statistics type to long

See merge request NGitLab!108

+++ NGitLab/NGitLab/Models/ProjectStatistics.cs
@@ -11,18 +11,18 @@ namespace NGitLab.Models
     public class ProjectStatistics
     {
         [DataMember(Name = "commit_count")]
-        public int CommitCount;
+        public long CommitCount;
 
         [DataMember(Name = "storage_size")]
-        public int StorageSize;
+        public long StorageSize;
 
         [DataMember(Name = "repository_size")]
-        public int RepositorySize;
+        public long RepositorySize;
 
         [DataMember(Name = "lfs_objects_size")]
-        public int LfsObjectsSize;
+        public long LfsObjectsSize;
 
         [DataMember(Name = "job_artifacts_size")]
-        public int JobArtifactsSize;
+        public long JobArtifactsSize;
     }
 }
